requirements.txt
Flask==3.0.0
flask-cors==4.0.0
alembic==1.13.1
requests==2.31.0
pytest==7.4.4
pytest-timeout==2.2.0
pytest-xdist==3.5.0
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
